refactor(Racing): 입력받은 game의 수에 대한 validate 처리 위임
