+++ src/test/java/racinggame/domain/racing/RacingTest.java
@@ -5,12 +5,14 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 class RacingTest {
@@ -30,6 +32,20 @@ class RacingTest {
     public void construct(int carCount, int gameCount, List<String> carNames) throws Exception {
         //then
         new Racing(carNames, gameCount);
+    }
+
+    @DisplayName("유효 하지 않은 게임 수 를 입력 하면 exception이 발생한다.")
+    @ParameterizedTest
+    @ValueSource(ints = {-50, -1, 0})
+    public void validGameCountInput(int count) throws Exception {
+        //given
+        final ArrayList<String> names = new ArrayList<>(Arrays.asList("a", "b", "c"));
+
+        //then
+        assertThatExceptionOfType(IllegalArgumentException.class)
+                .isThrownBy(
+                        () -> new Racing(names, count)
+                );
     }
 
     @DisplayName("레이싱 실행시 라운드수 와 레이싱 진행으로 인해 생긴 result의 크기를 비교")
